requirements.txt
mecab-python3==1.0.8
markovify==0.9.4
unidecode==unidecode-1.3.8